requirements.txt
Django==3.2
django-cleanup==5.2.0
django-generate-secret-key==1.0.2
djangorestframework==3.12.4
Pillow==8.2.0
audio-metadata==0.11.1
